+++ example/main.dart
@@ -18,11 +18,27 @@ void main(List<String> args) {
 
   print('abc@abc.test'.isEmailValid());
 
+  print('hello world'.encodeBase64());
+
+  print('aGVsbG8gd29ybGQ='.decodeBase64());
+
+  print('10'.isDigit());
+
+  print('lorem ipsum'.isLowerCase());
+
+  print('LOREM IPSUM'.isUpperCase());
+
+  print('ŹŻŚĄ å abcd'.isAlpha());
+
+  print('LorEM IpsUM'.swapCase());
+
   // Double
-  print(5.457.roundPrecision(1));
+  print(5.432.roundPrecision(1));
+
+  print(7.6.roundPrecision(0));
 
   // List
-  List<String> ls = ['A', 'B', 'C', 'D'];
+  List<String> ls = ['A', 'B', 'C', 'D', 'A', 'D', 'A'];
 
   ls.forEachIndexed((index, element) {
     print('Index $index - Element $element');
@@ -31,4 +47,6 @@ void main(List<String> args) {
   print(ls.random());
 
   print(ls.lastIndex);
+
+  print(ls.count((e) => e == 'A'));
 }