+++ lib/kt_utils.dart
@@ -2,7 +2,9 @@ library kt_utils;
 
 import 'dart:math';
 
-import 'package:kt_utils/regex_pattern.dart';
+import 'package:kt_utils/const/regex_pattern.dart';
+
+import 'dart:convert';
 
 part 'string_kt.dart';