+++ lib/list_kt.dart
@@ -17,4 +17,9 @@ extension ListKt<T> on Iterable<T> {
   T random() {
     return this.elementAt(Random().nextInt(this.length));
   }
+
+  ///Return count of the given item
+  int count(bool Function(T e) func) {
+    return this.where((element) => func(element)).length;
+  }
 }